+++ src/js/validator/remote.js
@@ -17,12 +17,15 @@
 
     FormValidation.Validator.remote = {
         html5Attributes: {
+            crossdomain: 'crossDomain',
+            data: 'data',
+            datatype: 'dataType',
+            delay: 'delay',
             message: 'message',
             name: 'name',
             type: 'type',
             url: 'url',
-            data: 'data',
-            delay: 'delay'
+            validkey: 'validKey'
         },
 
         /**
@@ -63,12 +66,10 @@
                 dfd.resolve($field, 'remote', { valid: true });
                 return dfd;
             }
-
-            var name    = $field.attr('data-' + ns + '-field'),
-                data    = options.data || {},
-                url     = options.url,
-                type    = options.type || 'GET',
-                headers = options.headers || {};
+            var name     = $field.attr('data-' + ns + '-field'),
+                data     = options.data || {},
+                url      = options.url,
+                validKey = options.validKey || 'valid';
 
             // Support dynamic data
             if ('function' === typeof data) {
@@ -86,18 +87,24 @@
             }
 
             data[options.name || name] = value;
+
+            var ajaxOptions = {
+                data: data,
+                dataType: options.dataType || 'json',
+                headers: options.headers || {},
+                type: options.type || 'GET',
+                url: url
+            };
+            if (options.crossDomain !== null) {
+                ajaxOptions.crossDomain = (options.crossDomain === true || options.crossDomain === 'true');
+            }
+
             function runCallback() {
-                var xhr = $.ajax({
-                    type: type,
-                    headers: headers,
-                    url: url,
-                    dataType: 'json',
-                    data: data
-                });
+                var xhr = $.ajax(ajaxOptions);
 
                 xhr
                     .success(function(response) {
-                        response.valid = response.valid === true || response.valid === 'true';
+                        response.valid = response[validKey] === true || response[validKey] === 'true';
                         dfd.resolve($field, 'remote', response);
                     })
                     .error(function(response) {